cant delete team with some match

+++ server/routes/team.js
@@ -1,5 +1,6 @@
 import { Router } from 'express'
 import { Team } from '../../schemas/team.js'
+import { Match } from '../../schemas/match.js'
 
 export const teamRouter = Router()
 
@@ -38,10 +39,29 @@ teamRouter.get('/:id', async (req, res) => {
   }
 })
 
+// teamRouter.delete('/:id', async (req, res) => {
+//   try {
+//     const id = req.params.id
+//     // Buscar el equipo por su ID y eliminarlo
+//     const deletedTeam = await Team.findByIdAndDelete(id)
+//     if (!deletedTeam) {
+//       return res.status(404).json({ message: 'No se encontró el equipo para eliminar' })
+//     }
+//     res.status(200).json({ message: 'Equipo eliminado exitosamente', deletedTeam })
+//   } catch (error) {
+//     console.error('Error al eliminar el equipo:', error)
+//     res.status(500).json({ message: 'Error al eliminar el equipo' })
+//   }
+// })
 teamRouter.delete('/:id', async (req, res) => {
   try {
     const id = req.params.id
-    // Buscar el equipo por su ID y eliminarlo
+    // Verificar si el equipo tiene algún partido asociado
+    const matchesCount = await Match.countDocuments({ $or: [{ homeTeam: id }, { awayTeam: id }] })
+    if (matchesCount > 0) {
+      return res.status(400).json({ message: 'No se puede eliminar el equipo porque tiene partidos asociados' })
+    }
+    // Si el equipo no tiene partidos asociados, proceder con la eliminación
     const deletedTeam = await Team.findByIdAndDelete(id)
     if (!deletedTeam) {
       return res.status(404).json({ message: 'No se encontró el equipo para eliminar' })
